+++ lib/screens/locations.dart
@@ -1,5 +1,5 @@
 import 'package:climate_app/screens/home.dart';
-import 'package:climate_app/services/api_openweather.dart';
+import 'package:climate_app/services/current_data_api.dart';
 import 'package:climate_app/widgets/weathercard.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
